requirements.txt
# Smart Surveillance System Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Computer Vision & Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-image==0.21.0

# Machine Learning & AI
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.0
face-recognition==1.3.0
mediapipe==0.10.7

# Biometric Processing
fingerprint-enhancer==0.1.0
deepface==0.0.79

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
redis==5.0.1

# Authentication & Security
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
cryptography==41.0.7

# Communication & Networking
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
paho-mqtt==1.6.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.10.1
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0